Init Project (#1)
This initial commit introduces the basic structure. The project include the model manager for user creation with related serializer and view for JWT tokens generation with user roles. Additionally, Candidate model has been linked to User via one-to-one relationship. The database setup for PostgreSQL and JWT configuration for token issuance have been added in settings.py. Dockerfile and docker-compose files have also been updated to setup the development environment with Django and PostgreSQL as a service.