Add private routes and candidate profile component
New private routes have been added to protect certain routes according to the authentication status of the user. A new component, CandidateProfile, was added to edit out a candidate's profile. Also created a user slice in Redux to manage the state of user profiles. Changes in the constants file were made to align the correct URL routes for API calls.