Update Header and Main components, add UserCard component
In Header.jsx, SIGNIN was replaced with SIGNUP and the rendering of sign in / sign up links was changed to account for different types of users - candidates or recruiters. This change allows the registration process to be more user-specific and tailored to each user type. In Main.jsx, unnecessary div tags were removed and remaining tags were restructured for cleaner code and improved readability. A new UserCard component was added to display user profiles more effectively.