Feat client candidate profile (#6)
* Update auth flow and component structure

Updated the SignIn and SignUp components, and authSlice.js for a more user-focused authentication flow. Improvements include differentiating between recruiter and candidate accounts, renaming tokens to accessToken and refreshToken for clarity, moved the Input component under Auth directory for better structure and replacing hardcoded 'email' value with the label prop in the Input Component.

* Add private routes and candidate profile component

New private routes have been added to protect certain routes according to the authentication status of the user. A new component, CandidateProfile, was added to edit out a candidate's profile. Also created a user slice in Redux to manage the state of user profiles. Changes in the constants file were made to align the correct URL routes for API calls.

* Update Header and Main components, add UserCard component

In Header.jsx, SIGNIN was replaced with SIGNUP and the rendering of sign in / sign up links was changed to account for different types of users - candidates or recruiters. This change allows the registration process to be more user-specific and tailored to each user type. In Main.jsx, unnecessary div tags were removed and remaining tags were restructured for cleaner code and improved readability. A new UserCard component was added to display user profiles more effectively.

* Extend JWT access token lifetime for testing purpose