Add role-based private routes and recruiter profile
This commit introduces a differentiation of private routes based on user roles. Instead of having one PrivateRoute component, there are two separate components CPrivateRoutes and RPrivateRoutes to handle route protection for candidate and recruiter roles respectively. This allows to have more control and customize the behaviour based on the user role.

This commit also includes creation of a new page RecruiterProfile. The profile page layout was previously blindly reusing CandidateProfile layout regardless of the role which was not suitable for recruiter role.