Feat client recruiter profile (#7)
* Add new components and refactor CandidateProfile.jsx

Added new components (Input, TextArea, CountrySelect) to improve the modularity and readability of the code. Refactored CandidateProfile.jsx to use these new components and implemented an asynchronous profile update function. The new function also handles errors and updates the Redux store with the user's changed name.

* Disable react prop-types linting rule

* Add handling of concurrent token refresh attempts

Mutex mechanism has been added to ensure that if multiple components request a new access token at the same time, they will have to wait for the ongoing refresh token request to finish before they proceed. Package 'async-mutex' was added to the project to manage the concurrency.

* Add JWT parsing and username handling to authSlice

* Add role-based private routes and recruiter profile

This commit introduces a differentiation of private routes based on user roles. Instead of having one PrivateRoute component, there are two separate components CPrivateRoutes and RPrivateRoutes to handle route protection for candidate and recruiter roles respectively. This allows to have more control and customize the behaviour based on the user role.

This commit also includes creation of a new page RecruiterProfile. The profile page layout was previously blindly reusing CandidateProfile layout regardless of the role which was not suitable for recruiter role.

* Replace user object with username in Header component

* Reduce JWT access token lifetime and allow nullable fields