Add sharedSlice for fetches and create CategorySelect component
In an effort to reduce duplication and to enhance the maintainability of the codebase, a sharedSlice.js file was created.

Additionally, components RecruiterProfile and CandidateProfile have been renamed to RecruiterOwnProfile and CandidateOwnProfile. This provides more clarity to the specific usage of these components - as they refer to the profiles of the current user.

In CandidateOwnProfile component, category input was replaced by CategorySelect component which fetches the options from the provided endpoint. This change ensures we have centralized, accurate, and dynamic categories data across the application.

Finally, a new file, CategorySelect.jsx, was created. This component fetches the list of all categories from the database and presents them to the users as selectable options.