Refactor UserCard, update navigation and profile components
Deleted UserCard component as it was replaced with UserCardGrid. This was done to align with the new design specs of displaying users in a grid rather than individually. Refactored navigation links to use NavLink instead of Link for appropriate styling. Also, adjusted profile related components and routes to specifically handle individual candidate and recruiter profiles for accurate display and functionality.