Feat user cards (#8)
* Add list candidate view and refactor serializers

* Update recruiter model and add new views and endpoints

Altered fields in the 'recruiter' model to allow 'about', 'city', and 'country' attributes to have null values. Added ListAPIView views for 'Category' and 'Skill' models and associated them with respective URL endpoints. Integrated 'getAllCandidates' query in the frontend 'userSlice'.

* Add sharedSlice for fetches and create CategorySelect component

In an effort to reduce duplication and to enhance the maintainability of the codebase, a sharedSlice.js file was created.

Additionally, components RecruiterProfile and CandidateProfile have been renamed to RecruiterOwnProfile and CandidateOwnProfile. This provides more clarity to the specific usage of these components - as they refer to the profiles of the current user.

In CandidateOwnProfile component, category input was replaced by CategorySelect component which fetches the options from the provided endpoint. This change ensures we have centralized, accurate, and dynamic categories data across the application.

Finally, a new file, CategorySelect.jsx, was created. This component fetches the list of all categories from the database and presents them to the users as selectable options.

* Refactor UserCard, update navigation and profile components

Deleted UserCard component as it was replaced with UserCardGrid. This was done to align with the new design specs of displaying users in a grid rather than individually. Refactored navigation links to use NavLink instead of Link for appropriate styling. Also, adjusted profile related components and routes to specifically handle individual candidate and recruiter profiles for accurate display and functionality.